# backend/src/main/resources/api-docs.yaml
openapi: 3.0.3
info:
  title: MyTrip API
  description: API documentation for MyTrip travel booking platform
  version: 1.0.0
  contact:
    name: MyTrip Support
    email: support@mytrip.com
servers:
  - url: http://localhost:8080/api
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /hotels/{id}:
    get:
      summary: Get hotel by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Hotel details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hotel'
        '404':
          description: Hotel not found
    put:
      summary: Update hotel
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Hotel'
      responses:
        '200':
          description: Hotel updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hotel'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (admin required)
        '404':
          description: Hotel not found
    delete:
      summary: Delete hotel
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Hotel deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (admin required)
        '404':
          description: Hotel not found
  /rooms:
    get:
      summary: Get all hotel rooms
      responses:
        '200':
          description: List of hotel rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HotelRoom'
    post:
      summary: Create a new hotel room
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HotelRoom'
      responses:
        '201':
          description: Hotel room created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HotelRoom'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (admin required)
  /rooms/{id}:
    get:
      summary: Get hotel room by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Hotel room details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HotelRoom'
        '404':
          description: Hotel room not found
    put:
      summary: Update hotel room
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HotelRoom'
      responses:
        '200':
          description: Hotel room updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HotelRoom'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (admin required)
        '404':
          description: Hotel room not found
    delete:
      summary: Delete hotel room
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Hotel room deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (admin required)
        '404':
          description: Hotel room not found
  /rooms/hotel/{hotelId}:
    get:
      summary: Get rooms by hotel ID
      parameters:
        - in: path
          name: hotelId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of rooms for the hotel
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HotelRoom'
        '404':
          description: Hotel not found
...
  /reviews/hotel/{hotelId}:
    get:
      summary: Get reviews for a hotel
      parameters:
        - in: path
          name: hotelId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of reviews for the hotel
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'
        '404':
          description: Hotel not found
...
  /notifications:
    get:
      summary: Get user's notifications
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of user's notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
        '401':
          description: Unauthorized
...
  /notifications/{id}/read:
    put:
      summary: Mark notification as read
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        '401':
          description: Unauthorized
        '404':
          description: Notification not found
  /notifications/{id}:
    delete:
      summary: Delete notification
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Notification deleted successfully
        '401':
          description: Unauthorized
        '404':
          description: Notification not found
  /price-history:
    get:
      summary: Get all price history records
      responses:
        '200':
          description: List of price history records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PriceHistory'
    post:
      summary: Create a new price history record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PriceHistory'
      responses:
        '201':
          description: Price history record created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceHistory'
  /price-history/{id}:
    get:
      summary: Get price history by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Price history record details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceHistory'
        '404':
          description: Price history record not found
  /price-history/hotel/{hotelId}:
    get:
      summary: Get price history for a hotel
      parameters:
        - in: path
          name: hotelId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Price history for the hotel
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PriceHistory'
        '404':
          description: Hotel not found
  /price-history/flight/{flightId}:
    get:
      summary: Get price history for a flight
      parameters:
        - in: path
          name: flightId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Price history for the flight
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PriceHistory'
        '404':
          description: Flight not found
...